backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-multipart==0.0.17
